Moved create product data to Context/Updated style of AddProduct page

+++ frontend-prod-list/src/components/UI/Navbar/Navbar.jsx
@@ -1,8 +1,9 @@
-import React from "react";
-import { Link } from "react-router-dom";
+import React, { Children } from "react";
+import { Link, useLocation } from "react-router-dom";
 import MyButton from "../button/MyButton";
 import { storeContext } from "../../../App";
 import axios from "axios";
+import BaseProductForm from "../../baseforms/BaseProductForm";
 
 const Navbar = () => {
   const {
@@ -12,51 +13,61 @@ const Navbar = () => {
     getProducts,
     loading,
     setLoading,
+    addNewProduct,
   } = React.useContext(storeContext);
 
   const deleteProduct = async () => {
-    console.log(booksType);
-    console.log(discsType);
-    console.log(furnitureType);
-
     const arr = [...booksType, ...discsType, ...furnitureType];
-    console.log(arr);
     const indexArr = [];
-    arr.map((element, index) => {
+    arr.map((element) => {
       if (element.isChecked) indexArr.push(element.id);
     });
 
-    // try {
-    //   const response = await axios.delete(`http://scandibackend/${indexArr}`);
-    //   console.log(response.data);
-    // } catch (err) {
-    //   alert(err);
-    // }
-    console.log(indexArr);
     try {
-      setLoading(true);
-      const response = await axios.delete(
-        `http://scandibackend/?id=${indexArr}`
-      );
-      console.log(response);
-      if (response.status === 200) {
-        getProducts();
+      if (indexArr.length > 0) {
+        setLoading(true);
+        const response = await axios.delete(
+          `http://scandibackend/?id=${indexArr}`
+        );
+        if (response.status === 200) {
+          getProducts();
+        }
+      } else {
+        alert(indexArr.length + " products was checked");
       }
     } catch (err) {
       alert(err);
     }
   };
 
+  const location = useLocation();
+  const isProductPage = location.pathname === "/products";
+
   return (
     <div className="navbar">
-      <Link to="/products">Product List</Link>
+      {location.pathname === "/products" ? (
+        <h2 style={{ color: "teal" }}>Product List</h2>
+      ) : (
+        <h2 style={{ color: "teal" }}>Product Add</h2>
+      )}
       <div className="navbar__links">
-        <Link to="/addproduct">
-          <MyButton>Add Product</MyButton>
-        </Link>
-        <MyButton style={{ background: "darkred" }} onClick={deleteProduct}>
-          Delete Product
-        </MyButton>
+        {location.pathname === "/products" ? (
+          <div>
+            <Link to="/addproduct">
+              <MyButton>Add Product</MyButton>
+            </Link>
+            <MyButton style={{ background: "darkred" }} onClick={deleteProduct}>
+              Delete Product
+            </MyButton>
+          </div>
+        ) : (
+          <div>
+            <MyButton onClick={addNewProduct}>Save</MyButton>
+            <Link to="/products">
+              <MyButton>Close</MyButton>
+            </Link>
+          </div>
+        )}
       </div>
     </div>
   );
